+++ src/components/app/details/cicdHistory/Artifacts.tsx
@@ -2,7 +2,6 @@ import React from 'react'
 import { showError, copyToClipboard } from '../../../common'
 import { toast } from 'react-toastify'
 import { useParams } from 'react-router'
-import { History } from '../cIDetails/types'
 import { ReactComponent as CopyIcon } from '../../../../assets/icons/ic-copy.svg'
 import { ReactComponent as Download } from '../../../../assets/icons/ic-download.svg'
 import { ReactComponent as MechanicalOperation } from '../../../../assets/img/ic-mechanical-operation.svg'
