+++ src/components/app/details/cicdHistory/Sidebar.tsx
@@ -8,7 +8,7 @@ import { STAGE_TYPE } from '../triggerView/types'
 import moment from 'moment'
 import { Moment12HourFormat, SourceTypeMap } from '../../../../config'
 import { CiPipelineSourceConfig } from '../../../ciPipeline/CiPipelineSourceConfig'
-import { CiMaterial, GitTriggers, History } from '../cIDetails/types'
+import { CiMaterial, GitTriggers, History } from './types'
 import TippyHeadless from '@tippyjs/react/headless'
 import { NavLink } from 'react-router-dom'
 import { statusColor as colorMap } from '../../config'
