+++ src/components/app/details/cicdHistory/TriggerDetails.tsx
@@ -2,14 +2,13 @@ import React, { useState } from 'react'
 import { Progressing, showError, createGitCommitUrl, asyncWrap, ConfirmationDialog } from '../../../common'
 import { toast } from 'react-toastify'
 import { useRouteMatch, useLocation, useParams } from 'react-router'
-import { History, GitTriggers, CiMaterial } from '../cIDetails/types'
 import { statusColor as colorMap } from '../../config'
 import { Moment12HourFormat } from '../../../../config'
 import moment from 'moment'
 import docker from '../../../../assets/icons/misc/docker.svg'
 import warn from '../../../../assets/icons/ic-warning.svg'
 import '../cIDetails/ciDetails.scss'
-import { PROGRESSING_STATUS, TERMINAL_STATUS_COLOR_CLASS_MAP } from '../cicdHistory/types'
+import { CiMaterial, GitTriggers, PROGRESSING_STATUS, TERMINAL_STATUS_COLOR_CLASS_MAP } from '../cicdHistory/types'
 import { Link } from 'react-router-dom'
 import { cancelCiTrigger, cancelPrePostCdTrigger } from '../../service'
