+++ src/components/app/service.ts
@@ -4,7 +4,7 @@ import { ResponseType } from '../../services/service.types'
 import { createGitCommitUrl, handleUTCTime, ISTTimeModal, sortCallback } from '../common'
 import moment from 'moment-timezone'
 import { ServerErrors } from '../../modals/commonTypes'
-import { History } from './details/cIDetails/types'
+import { History } from './details/cicdHistory/types'
 import { AppDetails, CreateAppLabelsRequest } from './types'
 import { CDMdalTabType, DeploymentWithConfigType } from './details/triggerView/types'
 import { AppMetaInfo } from './types'
